NUMBERS-100: Reduce scope of local variables
Several methods in the class org.apache.commons.numbers.fraction.FractionTest contain a number of tests that use shared local variables but are completely independent of each other because these local variables get assigned new values at the beginning of a test. Instead of using the same local variable throughout the method, the scope of these local variables is now reduced only to the test for which they are relevant.