Update Complex equals(Object) and hashCode().
The two methods are now equivalent to using java.util.Arrays to perform
the computation on an array of the real and imaginary components as if
using:

Arrays.equals(
  new double[]{c1.real(), c1.imaginary()},
  new double[]{c2.real(), c2.imaginary()},
  );

Arrays.hashCode(new double[]{real, imaginary});

This change does not alter the code of Complex.equals(Object) but
significantly updates the javadoc and tests for the equals method.

The hashCode() method has been updated to compute a different hash.
Tests have been extended to check the hash is computed as expected.