Optimise toString() and parse().
toString() may be longer than the default StringBuilder buffer size
(16). Use a larger value to avoid reallocation. This is set large enough
for 2 maximum precision doubles.

Use char as the format characters and not String. Requires changes to
the parsing. This avoids the use of String.split() which uses a regex
pattern.

Javadoc has been expanded taking style guides from the JDK Javadoc for
Double#toString() and Double.valueOf(String).