[NUMBERS-139] Implement multiply and divide for imaginary only.
multiplyImaginary(double)
divideImaginary(double)

Document that the method may return different signs than the
corresponding Complex method for the result parts if the complex has
zeros of the argument of the function is zero.

The documentation of the differences requires full unit test of all
cases. The exact cases that differ may be individually documented.