Approximation of sinh/cosh using exp can be overflow/underflow safe.
Updated the sinh(), cosh() and tanh() methods for large absolute real to
avoid overflow of exp(x) to infinite.

sinh()/cosh() now use a common function to compute the result as only
the sign of the result components differs. The computation of e^x can be
made in stages to allow computation of results even when e^x / 2
overflows.

tanh() now computes e^x in stages.

Added edge cases where exp(x) overflows but the result can be computed.