Change x2y2m1 summation algorithm.
This switches from Ogita's dot3 sum to Shewchuk's expansion sum. The
number of two sum operations is the same, the order is different.

The new version passes the previous strict tests but has a formal proof
in Shewchuk's paper. The method from Ogita may be functionally
equivalent but has no proof. Thus it may differ in some untested
combinations of x and y.