sqrt() to use multiply by sqrt(2) / 2 for imaginary only numbers.
The change alters the divide by sqrt(2) to a multiply by the closest
double representation of 1 / sqrt(2). This is sqrt(2) / 2.

This adds tests to ensure the polar representation is not used when
either a real or imaginary only number:

real = sqrt(abs()) * Math.cos(arg() / 2)
imag = sqrt(abs()) * Math.sin(arg() / 2)

Due to floating-point errors in the trignomic functions and the
inexactness of Math.PI the result is not the correct result. It
explicitly checks conditions that would fail if using the polar
computation with the trigonomic functions cos and sin.