Change tanh to use the identities for all computations.
This removes the requirement to assert that the switch from using
sin(2x) and cos(2x) and suitable identities is smooth. The test to
assert this failed on Java 11 but passed on Java 8. This may be due to a
change in the trigonomic functions between JDKs. The test assertion has
been removed as it is no longer valid for the new computation. The
functions to compute the identities for overflow are now redundant and
have been removed. The function is now overflow safe for finite
imaginary with no computation switch as y approaches infinity.

This lowers the ulp from the c reference test from 3 to 2 for tanh. Thus
the computation accuracy is similar and the c reference data may be
using the same computation.