Update attribution of 3rd party works in Complex
In accordance with https://www.apache.org/legal/src-headers.html#3party

- "Do not remove or modify any copyright notices or licenses"
- "Minor modifications/additions to 3rd party source files should
typically be licensed under the same terms as the rest of the 3rd party
source for convenience"

Here the 3rd party works are software conversions of single methods with
minor modifications. The methods are marked as distributed under the
original terms and the original copyright and license notice are
included inline.

Added copyright attribution to NOTICE.

Updated LICENSE to indicate the parts of 'commons-numbers' which are
subject to a different license.