Updated Fraction/BigFraction signum()
Use the method from BigInteger.divide(BigInteger).

The sign of the result is equal to a.signum() * b.signum() for a/b.

This exploits the fact that BigInteger explicitly stores the signum as a
field thus removing the need for signed comparators in the
BigFraction.signum() method.

For Fraction the use of Integer.signum(int) for numerator and
denominator is the equivalent operation. This is also branchless due to
use of bit manipulation in Integer.signum to compute the sign.