Change FractionException to use String.format
The use of MessageFormat.format does not support a canonical String
representation for a double value. This prevents the message from
containing the exact double that caused a conversion error.

For example the previous message format text contained the text:

Overflow trying to convert 0 to fraction

when the input double was 4.6566128730773926E-10