NUMBERS-147: Fix conversion from double to support 2^31
The value 2^31 can be in the numerator or denominator. Previously the
conversion from double supported up to Integer.MAX_VALUE which is
2^31-1.

Adds common test cases for conversion from double with a max
denominator.

Changed the conversion from a double to use the absolute value and
restore the sign at the end. The conversion is thus identical for
positive or negative values.

Throw an illegal argument exception if maxDenominator is zero. A zero
invalidates the conversion from a double with a restricted denominator.

Replace Math.abs with compareUnsigned. This is appropriate as the fraction p/q is
always positive.

Add test cases for overflow double conversions.

Fix the fall-back to always use p1/q1 when either p2 or q2 overflow.

Move the default max iterations to a constant.

Validate epsilon and max iterations are positive.