NUMBERS-147: Fix conversion from double to support 2^31
The value 2^31 can be in the numerator or denominator. Previously the
conversion from double supported up to Integer.MAX_VALUE which is
2^31-1.

Adds common test cases for conversion from double with a max
denominator.

Changed the conversion from a double to use the absolute value and
restore the sign at the end. The conversion is thus identical for
positive or negative values.