NUMBERS-183: Improve binomial coefficient classes
Refactor the tests to use JUnit 5 parameterized test features.

Refactor the test for all binomial coefficients up to n=200 to compute
the result using n! / (n-k)! / k!. Removes the caching test
implementation which only works for results within a long datatype. This
has been updated for BinomialCoefficientDouble and
LogBinomialCoefficient to test the result is correct even when a long
overflows.

Update the implementations:

- Avoid recursive method call if k > n/2.

- Add early exit if the result cannot fit into the output datatype. This
is k >= 34 for long, and k >= 515 for a double.

- BinomialCoefficientDouble: Use the precomputed factorials to compute
n! / k! / (n-k)! if possible.

- BinomialCoefficientDouble: avoid overflow by checking the intermediate
result

- LogBinomialCoefficient: Compute the terms using the LogBeta class