NUMBERS-190: Add a Bill of Materials module
Add a new module based on commons-rng-bom.

Configured maven-assembly-plugin skipAssembly in the parent pom as it
should only be configured in the dist-archive module. This allows the
snapshot BOM to be deployed to the snapshots repo using:

mvn deploy -Dgpg.skip

Deploy of the non-snapshot BOM will target the Apache release staging
repository and must be GPG signed.