Remove unused range arguments when creating the updating interval.
Note that the range is implied from the min and max index in the
indices.

The index out of bounds (IOOB) exceptions are not explicitly raised by
the internal methods. The public API will validate indices and array
ranges before calling the internal methods. The documentation of IOOB
exceptions by internal methods has been removed.